requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.2.2
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
Pillow==9.2.0
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.28.1
selenium==4.4.3
Selenium-Screenshot==2.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
webdriver-manager==3.8.3
Werkzeug==2.2.2
wsproto==1.1.0
zipp==3.8.1
